added simple tests for discounts

+++ ModelsBackend/models.py
@@ -153,7 +153,7 @@ class DiscountModel(models.Model):
     discountID = models.IntegerField(primary_key=True)
     category = models.CharField(null=True, max_length=100)
     productID = models.CharField(null=True, max_length=100)
-    percent = models.IntegerField()
+    percent = models.FloatField()
     type = models.CharField(max_length=100, choices=Simple_Discount_Choices)
     dID1 = models.ForeignKey('self', on_delete=models.CASCADE, related_name='firstDiscountID', null=True)
     dID2 = models.ForeignKey('self', on_delete=models.CASCADE, related_name='secondDiscountID', null=True)
